\name{multiQTL-package}
\alias{multiQTL-package}
\alias{multiQTL}
\docType{package}
\title{
A few plots to simplify extracting and plotting statistic from multiple QTL models
}
\description{
Integrates with R/qtl to improve extraction of statistics from multiple QTL models. Esspecially useful for multi-trait datsets. 
}
\details{
\tabular{ll}{
Package: \tab multiQTL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-19\cr
License: \tab NA \cr
}
}
\author{
John T. Lovell

Maintainer: John T. Lovell <johntlovell@gmail.com>
}
